pyproject.toml
[tool.poetry]
name = "liberty-gpt"
version = "0.1.0"
description = "work at home"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "liberty_gpt"}]

[tool.poetry.dependencies]
python = "^3.10.0"
langchain = "^0.0.227"
openai = "^0.27.8"
llama-index = "0.7.4"
streamlit = "1.24.0"
faiss-cpu = "^1.7.4"
python-dotenv = "^1.0.0"
transformers = "^4.30.2"
requests = "^2.31.0"
bs4 = "^0.0.1"
ipykernel = "^6.23.2"
backoff = "^2.2.1"
spacy = "3.5.4"
pypdf = "^3.10.0"
sentence-transformers = "^2.2.2"
sklearn = "^0.0.post5"
docx2txt = "^0.8"
python-magic-bin = "0.4.14"
# pycocotools-windows = "^2.0.0.2"
pypdf2 = "^3.0.1"
retry = "^0.9.2"
panel = "^1.1.1"
sqlalchemy = "^2.0.17"
bertopic = "^0.15.0"
networkx = "^3.1"
pyvis = "^0.3.2"
streamlit-option-menu = "^0.3.6"
fitz = "^0.0.1.dev2"
pymupdf = "^1.22.5"
html2text = "^2020.1.16"
markdownify = "^0.11.6"
pdfminer-six = "20200517"
sortedcontainers = "2.2.2"
reactpy = "^1.0.2"
langchain-decorators = "^0.0.11"
chromadb = "^0.3.26"
llama-hub = "^0.0.8"
chainlit = "^0.5.1"
unstructured = {extras = ["local-inference"], version = "^0.8.0"}
streamlit-chat = "^0.1.1"
duckduckgo-search = "^3.8.3"
tesseract = "^0.1.3"
poppler-utils = "^0.1.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.23.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
